Initial commit
This is intending to be a new network stack in C++. Maybe it will go
places, and maybe it will not.

Initial idea for structure is to have the class `packet` be the main
thing people pass around to move data through the system. Packet should
own a buffer of data that isn't moved for the duration of its existance.
Packets should be made up of headers that the programmer can step
through (`next_header`), and those should offer a view into the
information contained in each protocol layer (L2, L3, L4, etc).

This is super WiP and super uncertain at this point, I've barely started
implementing at this point.

Additionally, there are a few things of note about this project:
- This project contains both C and C++ code, and makes a very deliberate
distinction between headers that *are* C-compatible (`.h`) and those
that are *not* (`.hh`).
- I intend there to be some consistency in layout here, and thus new
files should copy `template.hh` as a template for layout.

Apart from that, let's explore and figure this out!